azure-mmr/requirements-api.txt
# requirements-api.txt - FastAPI backend dependencies
truststore>=0.9.1
azure-search-documents>=11.6.0
azure-storage-blob>=12.19.0
azure-ai-documentintelligence>=1.0.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
rich>=13.7.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0